add react-dom-route and its needed routes
